Split some DataGrid classes
